+++ src/ShaderPlayground.Core/Compilers/Mali/MaliCompiler.cs
@@ -15,14 +15,21 @@ namespace ShaderPlayground.Core.Compilers.Mali
 
             foreach (var versionDirectory in Directory.GetDirectories(Path.Combine(AppContext.BaseDirectory, "Binaries", "mali")))
             {
+                var maliscPath = Path.Combine(versionDirectory, "malisc.exe");
+                var maliocPath = Path.Combine(versionDirectory, "malioc.exe");
+
+                var isMalisc = File.Exists(maliscPath);
+
                 ProcessHelper.Run(
-                    Path.Combine(versionDirectory, "malisc.exe"),
+                    isMalisc ? maliscPath : maliocPath,
                     "--list",
                     out var stdOutput,
                     out var _);
 
                 // Extract cores from output.
-                var coreRegex = new Regex(@"\s+(Mali-[a-zA-Z0-9]+) <");
+                var coreRegex = isMalisc
+                    ? new Regex(@"\s+(Mali-[a-zA-Z0-9]+) <")
+                    : new Regex(@"(Mali-[a-zA-Z0-9]+) \(");
                 var matches = coreRegex.Matches(stdOutput);
 
                 coreOptions.AddRange(matches
@@ -48,30 +55,49 @@ namespace ShaderPlayground.Core.Compilers.Mali
         {
             CommonParameters.CreateVersionParameter("mali"),
             CommonParameters.GlslShaderStage,
-            CommonParameters.SpirVEntryPoint,
-            new ShaderCompilerParameter("Core", "Core", ShaderCompilerParameterType.ComboBox, CoreOptions, "Mali-G72")
+            new ShaderCompilerParameter("EntryPoint", "Entry point", ShaderCompilerParameterType.TextBox, defaultValue: "main", description: "Ignored when input language is GLSL.", filter: new ParameterFilter(CommonParameters.VersionParameterName, "6.2.0")),
+            new ShaderCompilerParameter("Core", "Core", ShaderCompilerParameterType.ComboBox, CoreOptions, "Mali-G78"),
+            new ShaderCompilerParameter("API", "API", ShaderCompilerParameterType.ComboBox, ApiOptions, "vulkan")
         };
 
         private static readonly string[] CoreOptions;
+
+        private static readonly string[] ApiOptions =
+        {
+            "opengles",
+            "vulkan",
+        };
 
         public ShaderCompilerResult Compile(ShaderCode shaderCode, ShaderCompilerArguments arguments)
         {
             var stage = GetStageFlag(arguments.GetString("ShaderStage"));
             var core = arguments.GetString("Core");
 
+            var maliscBinaryPath = CommonParameters.GetBinaryPath("mali", arguments, "malisc.exe");
+            var maliocBinaryPath = CommonParameters.GetBinaryPath("mali", arguments, "malioc.exe");
+            var isMalisc = File.Exists(maliscBinaryPath);
+
             var args = string.Empty;
             if (shaderCode.Language == LanguageNames.SpirV)
             {
-                var spirVEntryPoint = arguments.GetString("EntryPoint");
+                args += "--spirv";
 
-                args += "--spirv ";
-                args += $"--spirv_entrypoint_name {spirVEntryPoint}";
+                if (isMalisc)
+                {
+                    var spirVEntryPoint = arguments.GetString("EntryPoint");
+                    args += $" --spirv_entrypoint_name {spirVEntryPoint}";
+                }
+            }
+
+            if (!isMalisc)
+            {
+                args += $" --{arguments.GetString("API")}";
             }
 
             using (var tempFile = TempFile.FromShaderCode(shaderCode))
             {
                 ProcessHelper.Run(
-                    CommonParameters.GetBinaryPath("mali", arguments, "malisc.exe"),
+                    isMalisc ? maliscBinaryPath : maliocBinaryPath,
                     $"{stage} -c {core} {args} {tempFile.FilePath}",
                     out var stdOutput,
                     out var stdError);
@@ -79,8 +105,9 @@ namespace ShaderPlayground.Core.Compilers.Mali
                 return new ShaderCompilerResult(
                     true,
                     null,
-                    null, 
-                    new ShaderCompilerOutput("Output", null, stdOutput));
+                    !string.IsNullOrWhiteSpace(stdError) ? (int?)1 : null, 
+                    new ShaderCompilerOutput("Output", null, stdOutput),
+                    new ShaderCompilerOutput("Errors", null, stdError));
             }
         }